- Refactored ModifierConfigLogic.cs to ModifierConfigHandler.cs - Refactored Plugin.cs to have a dedicated method for initializing the config values for modifiers. - Added documentation to ModifierConfigHandler.cs & initialized configModifierWeights dictionary.
