+++ MonsterModifiers/MonsterModifiers.csproj
@@ -122,7 +122,7 @@
     <ItemGroup>
         <Compile Include="Managers\LocalizationManager.cs"/>
         <Compile Include="Properties\AssemblyInfo.cs"/>
-        <Compile Include="Src\Config\ModifierConfigLogic.cs" />
+        <Compile Include="Src\Config\ModifierConfigHandler.cs" />
         <Compile Include="Src\Custom Components\MonsterModifier.cs" />
         <Compile Include="Src\Modifiers\Absorption.cs" />
         <Compile Include="Src\Modifiers\BloodLoss.cs" />
